+++ src/main/java/com/Rest/CategoryProduct/pubsub/category/CategoryPubSubPublisher.java
@@ -1,17 +1,4 @@
 package com.Rest.CategoryProduct.pubsub.category;
-
-import com.google.cloud.spring.pubsub.core.PubSubTemplate;
-import org.springframework.stereotype.Service;
-
-@Service
-public class CategoryPubSubPublisher {
-    private final PubSubTemplate pubSubTemplate;
-
-    public CategoryPubSubPublisher(PubSubTemplate pubSubTemplate) {
-        this.pubSubTemplate = pubSubTemplate;
-    }
-
-    public void publish(String message){
-        pubSubTemplate.publish("yoj-category-topic", message);
-    }
+public interface CategoryPubSubPublisher {
+    public void publish(String message);
 }
